Rename spack.spack_context -> spack.context
