+++ lib/spack/spack/cmd/dependencies.py
@@ -11,7 +11,7 @@ from spack.active_environment import active_environment
 from spack.cmd.common import arguments
 from spack.solver.context import Context
 from spack.solver.input_analysis import create_graph_analyzer
-from spack.spack_context import SpackContext
+from spack.context import SpackContext
 from spack.util import tty
 from spack.util.tty.colify import colify
