+++ lib/spack/spack/context.py
