+++ lib/spack/spack/solver/reuse.py
@@ -17,7 +17,7 @@ from spack.active_environment import active_environment
 from spack.enums import InstallRecordStatus
 from spack.externals import ExternalSpecsParser
 from spack.externals_config import create_external_parser, external_config_with_implicit_externals
-from spack.spack_context import SpackContext
+from spack.context import SpackContext
 from spack.spec_filter import SpecFilter
 
 from .runtimes import all_libcs
